requirements.txt
django==5.2.8
djangorestframework==3.16.1
drf-yasg==1.21.11
celery==5.5.3
redis==7.0.1
psycopg2-binary==2.9.11
gunicorn==23.0.0


